+++ NINA.Core.WPF/ViewModel/AutoFocus/AutoFocusVM.cs
@@ -483,6 +483,37 @@ namespace NINA.WPF.Base.ViewModel.AutoFocus {
         }
 
         private async Task<bool> ValidateCalculatedFocusPosition(DataPoint focusPoint, FilterInfo filter, CancellationToken token, IProgress<ApplicationStatus> progress, double initialHFR) {
+            if (profileService.ActiveProfile.FocuserSettings.AutoFocusMethod == AFMethodEnum.STARHFR) {
+                // Evaluate R² for Fittings to be above threshold
+
+                var rSquaredThreshold = profileService.ActiveProfile.FocuserSettings.RSquaredThreshold;
+                if (rSquaredThreshold > 0) {
+                    var hyperbolicBad = HyperbolicFitting.RSquared < rSquaredThreshold;
+                    var quadraticBad = QuadraticFitting.RSquared < rSquaredThreshold;
+                    var trendlineBad = TrendlineFitting.LeftTrend.RSquared < rSquaredThreshold || TrendlineFitting.RightTrend.RSquared < rSquaredThreshold;
+
+                    var fitting = profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting;
+
+                    if ((fitting == AFCurveFittingEnum.HYPERBOLIC || fitting == AFCurveFittingEnum.TRENDHYPERBOLIC) && hyperbolicBad) {
+                        Logger.Error($"R² (Coefficient of determination) for Hyperbolic Fitting is below threshold. {Math.Round(HyperbolicFitting.RSquared, 2)} / {rSquaredThreshold}");
+                        Notification.ShowError(string.Format(Loc.Instance["LblAutoFocusCurveCorrelationCoefficientLow"], Math.Round(HyperbolicFitting.RSquared, 2), rSquaredThreshold));
+                        return false;
+                    }
+
+                    if ((fitting == AFCurveFittingEnum.PARABOLIC || fitting == AFCurveFittingEnum.TRENDPARABOLIC) && quadraticBad) {
+                        Logger.Error($"R² (Coefficient of determination) for Parabolic Fitting is below threshold. {Math.Round(QuadraticFitting.RSquared, 2)} / {rSquaredThreshold}");
+                        Notification.ShowError(string.Format(Loc.Instance["LblAutoFocusCurveCorrelationCoefficientLow"], Math.Round(QuadraticFitting.RSquared, 2), rSquaredThreshold));
+                        return false;
+                    }
+
+                    if ((fitting == AFCurveFittingEnum.TRENDLINES || fitting == AFCurveFittingEnum.TRENDHYPERBOLIC || fitting == AFCurveFittingEnum.TRENDPARABOLIC) && trendlineBad) {
+                        Logger.Error($"R² (Coefficient of determination) for Trendline Fitting is below threshold. Left: {Math.Round(TrendlineFitting.LeftTrend.RSquared, 2)} / {rSquaredThreshold}; Right: {Math.Round(TrendlineFitting.RightTrend.RSquared, 2)} / {rSquaredThreshold}");
+                        Notification.ShowError(string.Format(Loc.Instance["LblAutoFocusCurveCorrelationCoefficientLow"], Math.Round(TrendlineFitting.LeftTrend.RSquared, 2), Math.Round(TrendlineFitting.RightTrend.RSquared, 2), rSquaredThreshold));
+                        return false;
+                    }
+                }
+            }
+
             var min = FocusPoints.Min(x => x.X);
             var max = FocusPoints.Max(x => x.X);